fix directories
