--- src/Home.jsx
@@ -1,10 +1,0 @@
-import Carrousel from "./components/carrousel/Carrousel.jsx";
-
-export const Home = () => {
-  return (
-    <>
-          <h2>Soy la página de inicio</h2>
-          <Carrousel/>
-    </>
-  );
-};
