+++ src/LandingPage/Home/Home.jsx
@@ -1,0 +1,10 @@
+import Carrousel from "../Carrousel/Carrousel";
+
+export const Home = () => {
+  return (
+    <>
+      <h2>Soy la página de inicio</h2>
+      <Carrousel />
+    </>
+  );
+};
