+++ src/ProtectedRoute/ProtectedRoute.jsx
@@ -1,7 +1,5 @@
-import { useAuth } from "../../Context/authContext";
+import { useAuth } from "../Context/authContext";
 import { Navigate, useNavigate } from "react-router-dom";
-import Dashboard from "../Dashboard.jsx";
-import { UserView } from "../userView/UserView";
 
 // Garantiza la proteccion de la ruta para que nadie que no esté logueado lo vea
 export const ProtectedRoute = ({ children }) => {
