+++ src/ProtectedRoute/ProtectedRouteUser.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { useAuth } from "../Context/authContext";
 import { Navigate } from "react-router-dom";
-import { userView } from "./userView/UserView";
 
 const ProtectedRouteUser = ({ children }) => {
   const { user, users, loading } = useAuth();
