--- src/components/User.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const User = () => {
-    return <>
-        <h1>Soy el contenido del usuario</h1>
-    </>
-}
